# Prometheus: reduce memory allocations in readLabelsOrExemplars.